[iOS] refactor: TokenAddCellView 분리
