+++ iOS/DaDaIkSeon/DaDaIkSeon/Main/View/TokenCellView.swift
@@ -62,40 +62,3 @@ struct TokenCellView: View {
         .shadow(color: Color.shadow, radius: 6, x: 0, y: 3.0)
     }
 }
-
-// MARK: TokenAddCellView
-
-struct TokenAddCellView: View {
-    
-    // MARK: Body
-    
-    @State var showSheetView = false
-    
-    var body: some View {
-        
-        //        NavigationLink(
-        //            destination: QRGuideView(),
-        //            label: {
-        VStack {
-            Spacer()
-            HStack {
-                Spacer()
-                Image(systemName: "plus.circle")
-                    .resizable()
-                    .frame(width: 25, height: 25, alignment: .center)
-                    .foregroundColor(Color(.systemGray2))
-                
-                Spacer()
-            }
-            Spacer()
-        }
-        .padding(.horizontal, 12)
-        .overlay(
-            RoundedRectangle(cornerRadius: 15)
-                .strokeBorder(style: StrokeStyle(lineWidth: 2, dash: [5.0]))
-                .foregroundColor(Color(.systemGray2))
-        )
-        //            })
-    }
-}
-
